requirements.txt
flask>=3.0.0
flask-limiter>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
python-dotenv>=1.0.0